Language, 404 page, style, flaggs, logos.
Changing language now does not resets the page.

Added new buttons in the navbar to change the language.

button of the language that was not selected grayed out.

Added new English translations to the contact form.

Added a new shiny 404 page. (with translations)

Fixed some typos.

Added css for the contact form.

Added the footer images.